pyproject.toml
[tool.poetry]
name = "protect_exporter"
version = "0.2.0"
description = "Export UniFi Protect metrics to Prometheus"
authors = ["Jinna Kiisuo <jinna+git@nocturnal.fi>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8.1"
absl-py = "^2.0.0"
prometheus-client = "^0.13.0"
asyncio = "^3.4.3"
aiohttp = "^3.9.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"
flake8 = "^6.1.0"
black = "^23.7.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
protect_exporter = 'protect_exporter:run'

[tool.mypy]
follow_imports = "silent"
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "absl"
ignore_missing_imports = true
[[tool.mypy.overrides]]
module = "prometheus_client"
ignore_missing_imports = true
